refactoring process
